Reuse tools from optimade-python-tools (#39)
Use exception handlers and config from optimade.
Add mongo from optimade and set to min. v0.3.2.
