+++ aiida_optimade/config.json
@@ -1,12 +1,13 @@
 {
   "page_limit": 15,
-  "db_page_limit": 500,
-  "index_base_url": null,
+  "page_limit_max": 500,
+  "base_url": null,
   "provider": {
     "prefix": "_aiida_",
     "name": "AiiDA",
     "description": "AiiDA: Automated Interactive Infrastructure and Database for Computational Science (http://www.aiida.net)",
-    "homepage": "http://www.aiida.net"
+    "homepage": "http://www.aiida.net",
+    "index_base_url": null
   },
   "implementation": {
     "name": "aiida-optimade",
@@ -17,7 +18,6 @@
   "provider_fields": {
     "structures": [
       "ctime"
-    ],
-    "references": []
+    ]
   }
 }
